pyproject.toml
[tool.poetry]
name = "optimality-of-decentralization"
version = "0.1.0"
description = ""
authors = ["Prateek Bhustali <p.bhustali@tudelft.nl>"]
readme = "README.md"
packages = [{ include = "imprl", from = "imprl/src" }]

[tool.poetry.dependencies]
python="~3.9"
numpy="^1.26.2"
scipy="^1.12.0"
torch="~2.2.1"
pyyaml="~6.0.1"
gym="^0.23.1"
notebook="^7.1.2"
rliable="~1.0.8"
"ruamel.yaml"="~0.18.6"

[tool.poetry.group.logging]
optional = false
[tool.poetry.group.logging.dependencies]
wandb="~0.16.4"

[tool.poetry.group.visualisation]
optional = false
[tool.poetry.group.visualisation.dependencies]
matplotlib="^3.8.2"
seaborn="^0.13.2"

[tool.poetry.group.dev]
optional = true
[tool.poetry.group.dev.dependencies]
pytest="~8.0.1"
black="^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
